feat: 从备份恢复完整代码(含登录界面) + versionCode 148 + 修复location-sync字段名(userId/elderId)

+++ app/src/main/java/com/falldetector/diedaobao/cloud/WSClient.kt
@@ -276,7 +276,7 @@ object WSClient {
             put("data", JSONObject().apply {
                 put("latitude", latitude)
                 put("longitude", longitude)
-                put("accuracy", accuracy)
+                put("accuracy", accuracy.toDouble())  // Fix: Float -> Double
                 put("timestamp", System.currentTimeMillis())
             })
         }
